% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAverage.R
\name{calcAverage}
\alias{calcAverage}
\title{calcAverage}
\usage{
calcAverage(df, thresh, rolling_spc)
}
\arguments{
\item{df}{data.frame prepped by \code{parse_exceedance()}}

\item{thresh}{air quality threshold definitions}

\item{rolling_spc}{names of species for which \code{calcRolling()} should be called. Others are passed to \code{calcAverage()}}
}
\description{
Calculates rolling or fixed averaging periods for \code{parse_exceedance()}
}
\author{
W. S. Drysdale
}
