% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFixed.R
\name{calcFixed}
\alias{calcFixed}
\title{calcFixed}
\usage{
calcFixed(df, thresh)
}
\arguments{
\item{df}{data.frame prepped by \code{parse_exceedance()}}

\item{thresh}{air quality threshold definitions}
}
\description{
Calculates means over fixed periods for \code{parse_exceedance()}. Periods are calcualted via \code{hour_grouping()}
}
\author{
W. S. Drysdale
}
