% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRolling.R
\name{calcRolling}
\alias{calcRolling}
\title{calcRolling}
\usage{
calcRolling(df, thresh)
}
\arguments{
\item{df}{data.frame prepped by \code{parse_exceedance()}}

\item{thresh}{air quality threshold definitions}
}
\description{
Calculates rolling means for \code{parse_exceedance()}
}
\author{
W. S. Drysdale
}
