% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_exceedance.R
\name{count_exceedance}
\alias{count_exceedance}
\title{count_exceedance}
\usage{
count_exceedance(
  df,
  thresh,
  period = 1,
  exceedance_days_only = "o3_8_100",
  summary = c("none", "total", "average")[1],
  summary_name = NULL,
  total = FALSE,
  total_name = NULL
)
}
\arguments{
\item{df}{output of \code{parse_exceedance()}}

\item{thresh}{air quality threshold definitions}

\item{period}{interger or vector of interger periods to count. default c(1,8,24) i.e, count hourly, 8 hourly and daily exceednces defined in thresh}

\item{exceedance_days_only}{names of threshold groups that where multiple exceedances in one calenday day only count as one. \cr
useful to remove overcounting of rolling thresholds}

\item{summary}{one of c("none", "total", "average"). 
               \itemize{
                \item none - count per code
                \item total - sum across all codes
                \item average - average per code
               }
None returns exceedances per group defined by the code column}

\item{summary_name}{if using total or average name the new code grouping}
}
\description{
Count the exceedances, grouped by code for species defined in name. Operated on the output of \code{parse_exceedance()}
}
\author{
W. S. Drysdale
}
