% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_exceedance_plotList.R
\name{create_exceedance_plotList}
\alias{create_exceedance_plotList}
\title{create_exceedance_plotList}
\usage{
create_exceedance_plotList(count_df, spc = c("no2_1_200", "pm10_24_50",
  "pm25_24_25"), max_sites = 9, year_range = 2019:2020,
  top_year = 2019)
}
\arguments{
\item{count_df}{output of \code{count_exceedance()}}

\item{spc}{character vector of speices to be plotted in threshold format}

\item{max_sites}{maximum number of sites to show for each spc. only for summary = "none" of \code{count_exceedance}. Default 9}

\item{year_range}{integer vector of years to show exceedances for, defaults 2019:2020}

\item{top_year}{which year should be used to order the sites by number of exceedances. Default 2019}
}
\description{
Wrapper around \code{plot_exceedance()} to allow more control over the resulting plots
}
\author{
W. S. Drysdale
}
