% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aurn.R
\name{get_aurn}
\alias{get_aurn}
\title{get_aurn}
\usage{
get_aurn(keyword = NULL, site.type = NULL, meta = NULL, ...)
}
\arguments{
\item{keyword}{Optional keyword matched to site name using \code{stringr::str_detect()}}

\item{site.type}{One, or a vector of: "Urban Background", "Urban Traffic", "Rural Background", "Suburban Background", "Urban Industrial", "Suburban Industrial". NULL imports all site types}

\item{meta}{Optionally supply the site meta data. When NULL this is automatically downloaded via \code{openair::importMeta(source = "aurn")}}

\item{...}{additonal paramerters passed to \code{openair::importMeta()}}
}
\description{
Wrapper around \code{openair::importAURN()} to download and shape AURN data
}
\author{
W. S. Drysdale
}
