% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_sites.R
\name{get_top_sites}
\alias{get_top_sites}
\title{get_top_sites}
\usage{
get_top_sites(df, year = 2019)
}
\arguments{
\item{df}{\code{ount_exceedance} output}

\item{year}{the year to sort exceedances by}
}
\description{
helper to organise sites into those with the most exceedances, used in \code{create_exceedance_plotList}
}
\author{
W. S. Drysdale
}
