% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hour_grouping.R
\name{hour_grouping}
\alias{hour_grouping}
\title{hour_grouping}
\usage{
hour_grouping(period)
}
\arguments{
\item{period}{number of hours in grouping period}
}
\description{
Helper function which returns groupings for fixed averaging intervals over 24 hours. Only supports periods that are a factor of 24
}
\author{
W. S. Drysdale
}
