% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_AQ_stripes.R
\name{parse_AQ_stripes}
\alias{parse_AQ_stripes}
\title{parse_AQ_stripes}
\usage{
parse_AQ_stripes(df, type = c("conc", "diff")[1], preserve_code = F)
}
\arguments{
\item{df}{data.frame of one site's data}

\item{type}{one of c("conc","diff").
\itemize{
 \item conc - format concentrations as supplied
 \item diff - calcualte the difference between the daily median of teh frist n-1 years supplied vs the nth year
 }}

\item{preserve_code}{Logical. Restore the code column to the output}
}
\description{
Parse data to produce air quality stripes
}
\author{
W. S. Drysdale
}
