% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_AQ_stripes_sitegrid.R
\name{parse_AQ_stripes_sitegrid}
\alias{parse_AQ_stripes_sitegrid}
\title{parse_AQ_stripes_sitegrid}
\usage{
parse_AQ_stripes_sitegrid(siteList, type = c("conc", "diff")[1],
  year = 2020)
}
\arguments{
\item{siteList}{list of individual site data}

\item{type}{one of c("conc","diff").
\itemize{
 \item conc - format concentrations as supplied
 \item diff - calcualte the difference between the daily median of teh frist n-1 years supplied vs the nth year
 }}

\item{year}{optional numeric vector. Filter the output data by year (more input data than needs to be plotted may be required for \code{type == "diff")}}
}
\description{
Wrapper around \code{parse_AQ_stripes()}. Produces data to plot concentraions on a site by site basis
}
\author{
W. S. Drysdale
}
