% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_exceedance.R
\name{parse_exceedance}
\alias{parse_exceedance}
\title{parse_exceedance}
\usage{
parse_exceedance(dat, thresh, rolling_spc = "o3")
}
\arguments{
\item{dat}{data.frame of site(s) data}

\item{thresh}{air quality threshold definitions}

\item{rolling_spc}{names of species for which \code{calcRolling()} should be called. Others are passed to \code{calcAverage()}}
}
\description{
Parse data so air quality threshold exceedances can be counted.
}
\author{
W. S. Drysdale
}
