% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_spec.R
\name{parse_spec}
\alias{parse_spec}
\title{parse_spec}
\usage{
parse_spec(df, name = name, unit = TRUE, TeX = TRUE)
}
\arguments{
\item{df}{data.frame containing a column of species names, named "name" by default}

\item{name}{name of column containing species names to be parsed. Supplied unquoted}

\item{unit}{logical, should the unit be included in the parsed output? Default TRUE}

\item{TeX}{logical, should string be wrapped in \code{paste0(latex2exp::TeX())}? Default TRUE}
}
\description{
Helper to format common species for plotting. Adds a new column called \code{name_parsed}
}
\author{
W. S. Drysdale
}
