% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_year_comp.R
\name{parse_year_comp}
\alias{parse_year_comp}
\title{parse_year_comp}
\usage{
parse_year_comp(df, comp_year = 2020, species = c("no2", "pm25", "o3"),
  rolling_width = 15, grp = NULL, by_code = F, city = NULL)
}
\arguments{
\item{df}{data.frame of site data}

\item{comp_year}{what year should be compared, note currently only year previous to this will be compared}

\item{species}{names of speices to perform comparison on}

\item{rolling_width}{width of the rolling median in days}

\item{grp}{optional grouping label - it is needed less now the function supports grouping by code}

\item{by_code}{logical, should the function be applied to each code group individually? default false}
}
\description{
Compare a year's timeseries to previous years.
}
\author{
W. S. Drysdale
}
