% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_AQ_stripes.R
\name{plot_AQ_stripes}
\alias{plot_AQ_stripes}
\title{plot_AQ_stripes}
\usage{
plot_AQ_stripes(df, spc = "no2", vline_pos = NULL, title = NULL)
}
\arguments{
\item{df}{output of \code{parse_AQ_stripes}}

\item{spc}{name of species in name column to plot}

\item{vline_pos}{numeric vector. add vertical lines at the given day of year to stripes of the nth year}

\item{title}{add a title to the plot via \code{ggtitle()}}
}
\description{
Plot air quality stripes
}
\author{
W. S. Drysdale
}
