% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_AQ_stripes_sitegrid.R
\name{plot_AQ_stripes_sitegrid}
\alias{plot_AQ_stripes_sitegrid}
\title{plot_AQ_stripes_sitegrid}
\usage{
plot_AQ_stripes_sitegrid(
  stripeDat,
  spc = "no2",
  yday_thresh = (yday(Sys.time()) - 1),
  normalise = F
)
}
\arguments{
\item{stripeDat}{output of \code{parse_AQ_stripes_grid()}}

\item{spc}{name of species in name column to plot}

\item{yday_thresh}{option cutoff for x axis. Default day of year of system time, less 1}

\item{normalise}{logical. normalise each sites data. For conc between 0 and 1, diff -1 and 1}
}
\description{
Plot air quality stripes as a grid by site
}
\author{
W. S. Drysdale
}
