% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exceedance.R
\name{plot_exceedance}
\alias{plot_exceedance}
\title{plot_exceedance}
\usage{
plot_exceedance(count_df)
}
\arguments{
\item{count_df}{output of \code{count_exceedance()}}
}
\description{
Plots the output of \code{count_exceedance()}. Is wrapped by \code{create_exceedance_plotList}) for more control
}
\author{
W. S. Drysdale
}
