% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseSitesByDateRange.R
\name{summariseSitesByDateRange}
\alias{summariseSitesByDateRange}
\title{summariseSitesByDateRange}
\usage{
summariseSitesByDateRange(
  df,
  date_start = lubridate::ymd_hm("2020-01-01 00:00"),
  date_end = lubridate::ymd_hm("2020-01-31 00:00"),
  by_code = F
)
}
\arguments{
\item{df}{data frame containing date, value/diff and optionally code}

\item{date_start}{POSIXct refering to the beginning of the period to be summarised}

\item{date_end}{POSIXct refering to the end of the period to be summarised}

\item{by_code}{logical, should the data be grouped by code}
}
\description{
Return mean median and standard deviation of value or diff column
}
\author{
W. S. Drysdale
}
