% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_aurn.R
\name{tidy_aurn}
\alias{tidy_aurn}
\title{Tidy Aurn}
\usage{
tidy_aurn(dat)
}
\arguments{
\item{dat}{output of \code{openair::import_aurn()}}
}
\description{
Helper function to convert \code{openair::import_aurn()} output into a format \code{AQVisR} functions expect
}
\author{
W. S. Drysdale
}
