% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_week_bar.R
\name{plot_week_bar}
\alias{plot_week_bar}
\title{Plot week bar}
\usage{
plot_week_bar(df, current = "2020", previous = "2015 - 2019",
  by_code = T)
}
\arguments{
\item{df}{output of \code{parse_year_comp()}}

\item{current}{name of group in column "y" to be subtracted from. default 2020}

\item{previous}{name of group in column "y" to subtract from current, default "2015 - 2019"}

\item{by_code}{logical, should the analysis be grouped by the code column}
}
\description{
Plots the output of \code{parse_year_comp()} as a bar chart of the difference of two periods
}
\author{
W. S. Drysdale
}
