% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_week_bar.R
\name{plot_week_bar}
\alias{plot_week_bar}
\title{Plot week bar}
\usage{
plot_week_bar(
  df,
  current = "2020",
  previous = "2015 - 2019",
  by_code = T,
  vline_pos = NULL,
  xlim = c(0, 52),
  na.rm = F,
  type = c("difference", "percent", "absolute")[1],
  combine_spc = F,
  highlight_range = xlim
)
}
\arguments{
\item{df}{output of \code{parse_year_comp()}}

\item{current}{name of group in column "y" to be subtracted from. default 2020}

\item{previous}{name of group in column "y" to subtract from current, default "2015 - 2019"}

\item{by_code}{logical, should the analysis be grouped by the code column}

\item{vline_pos}{numeric vector to place vertical lines}

\item{na.rm}{logical, should the data be filtered by \code{filter(!is.na(value))}}

\item{type}{one of "difference", "percent" or "absolute".
\itemize{
    \item {difference - change is current-previous}
    \item {percent - change is ((current-previous)/previous)*100}
    \item {absolute - plot absolute values from current period only}
}}

\item{combine_spc}{logical, should species be used in facet wrapping}

\item{highlight_range}{numerical vector length 2. range of values to highlight.
bars that fall outside of this range have their alpha reduced. default equal to xlim}
}
\description{
Plots the output of \code{parse_year_comp()} as a bar chart of the difference of two periods
}
\author{
W. S. Drysdale
}
